Move dashboard code into different module
